refactor: cjs to mjs
